finish image upload api, return string, and init image analyzer and file handler
