+++ app/config.py
@@ -1,4 +1,4 @@
 class Config:
     SECRET_KEY = "your_secret_key"  # 替換為安全的密鑰
-    UPLOAD_FOLDER = "uploads"
+    UPLOAD_FOLDER = "app/uploads"
     ALLOWED_EXTENSIONS = {"jpg", "jpeg", "png"}
