--- app/route/inference.py
@@ -1,6 +1,0 @@
-def predict_disease(image_path):
-    # 假設這裡使用一個預載模型
-    # TODO: 加入模型載入和推論邏輯
-    # 示例: 模型讀取圖片並返回結果
-    prediction = {"disease": "example_disease", "confidence": 0.95}
-    return prediction
