+++ app/utils/image_analyzer.py
@@ -1,0 +1,7 @@
+def analyze_image(filepath):
+    """
+    模擬圖片分析邏輯
+    TODO: 替換為您的實際圖片處理邏輯（例如調用機器學習模型）
+    """
+    # 示例：假設處理後返回固定文本
+    return "This is a simulated analysis result for the uploaded image."
